python/pyproject.toml
[tool.poetry]
name = "bos-api-demo"
version = "0.1.0"
description = ""
authors = ["Braiins Systems s.r.o <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.38.1"
grpcio-tools = "^1.38.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

